fix: the charge-up lights the piece from within, and bishops aim at White
**The charge-up looked like check.** It was a pulsing ring behind the sprite;
so is the check halo, at a slightly different radius and colour. Ring vocabulary
is spoken for three times over — check, the king's shield, deflection — so the
cue is now a tinted additive copy of the piece's own texture laid over it. What
brightens is the silhouette itself: no new shape appears, and it cannot be
confused with anything else on the board. It swells 6% as it burns.

**The king's shield is tighter and quieter.** r 0.42 → 0.33 (it was wider than
the check halo and read as a circle drawn on the board rather than a field the
piece is wearing), line 1 → 0.8, glow 4 → 2.5, fill and pulse both dialled back.
`flareForcefield` still spikes it on every absorbed hit, which is when it should
be visible; at rest it should be easy to overlook.

**Bishops aim.** A fixed 45° from the back rank is useless: the round crosses
seven files before it reaches White and leaves the board entirely. Measured
across every shooter position, a third of them threw the shot away. The lean is
now taken from an actual white piece, picked fresh each beat, and clamped to
17°–45° — shallower reads as a straight shot that missed, steeper is back to
flying off the side. Same measurement afterwards: 4%.

That made `lean` a slope rather than -1/0/+1, which the projectile geometry
already handled — `atan2(dx, -dy)` and the path-length duration were both
derived from the travel vector, so they generalise for free. The telegraph tick
uses the same formula, so the cue cannot promise one angle and the missile take
another.

+++ GalacticChessInvaders/Game/Rendering/LaserNode.swift
@@ -145,12 +145,13 @@ final class LaserNode: SKSpriteNode {
     /// Fires straight up (player, from the ship) or down (enemy, from a fleet
     /// piece) from `origin`, dealing `state.damage` on contact. Deactivates on
     /// its own after `travelDistance` if nothing hits it first.
-    /// `lean` angles the shot: 0 straight, -1 left, +1 right. A leaning shot
-    /// travels 45°, so it covers `travelDistance` on both axes and therefore a
-    /// sqrt(2) longer path — the duration accounts for that, or an angled shot
-    /// would cross the board faster than its stated speed (§21.3).
+    /// `lean` is the slope: sideways travel per unit of forward travel, 0 for
+    /// straight, negative left. An angled shot therefore covers a longer path
+    /// than `travelDistance`, and the duration is taken from the real path
+    /// length — otherwise an angled round would cross the board faster than its
+    /// stated speed (§21.3).
     func fire(from origin: CGPoint, damage: Int, speed: CGFloat,
-              travelDistance: CGFloat, lean: Int = 0) {
+              travelDistance: CGFloat, lean: CGFloat = 0) {
         guard speed > 0, travelDistance > 0 else { return }
         state = ProjectileState(owner: owner, damage: damage, speed: speed)
         position = origin
@@ -160,7 +161,7 @@ final class LaserNode: SKSpriteNode {
         physicsBody?.contactTestBitMask = liveContactMask
 
         let dy = owner == .player ? travelDistance : -travelDistance
-        let dx = CGFloat(lean) * travelDistance
+        let dx = lean * travelDistance
         let path = (dx * dx + dy * dy).squareRoot()
         let duration = TimeInterval(path / speed)
